+++ src/Assets.php
@@ -79,7 +79,6 @@ class Assets
     public function get(AssetType|string $type, string $namespace = self::NAMESPACE_COMMON): string
     {
         $type = AssetType::normalize($type);
-
         $paths = $this->getResults($type, $this->assetsCollection->get($type, $namespace));
 //        return RenderFactory::getRender(\strtolower($type), $this->environment)->getResult($paths);
         return $this->getEnvironment()->getRenderer($type)->render($paths);
@@ -87,14 +86,12 @@ class Assets
 
 
     /**
-     * @param AssetType|string $type
+     * @param AssetType $type
      * @param array<Asset> $assetsCollection
      * @return array
      */
-    private function getResults(AssetType|string $type, array $assetsCollection): array
+    private function getResults(AssetType $type, array $assetsCollection): array
     {
-        $type = AssetType::normalize($type);
-
         $strategy = StrategyFactory::getStrategy(
             $this->environment,
             $assetsCollection,
